requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
flake8
python-dotenv
